Break up tests and remove interdependency until socrata-test-common PR lands
