+++ balboa-agent/src/test/scala/com/socrata/balboa/agent/HttpMetricQueueSpec.scala
@@ -2,6 +2,7 @@ package com.socrata.balboa.agent
 
 import java.net.URL
 
+import com.fasterxml.jackson.core.JsonParseException
 import com.socrata.balboa.metrics.Metric.RecordType
 import com.socrata.metrics.{DomainId, MetricIdParts, UserUid, ViewUid}
 import com.stackmob.newman.{ApacheHttpClient, Headers, RawBody}
@@ -12,10 +13,12 @@ import org.mockito.Mockito._
 import org.mockito.Matchers._
 import org.scalatest.mock.MockitoSugar
 
-import com.socrata.balboa.util.TestUtil._
+import org.json4s._
+import org.json4s.jackson.JsonMethods.{parse, pretty, render}
 
 import scala.concurrent.Future
 import scala.concurrent.duration._
+import scala.util.Try
 
 /**
   * Unit Tests for [[MetricConsumer]]
@@ -25,6 +28,24 @@ class HttpMetricQueueSpec extends WordSpec
   with ShouldMatchers
   with MockitoSugar
   with StrictLogging {
+
+  // To remove once PR #6 lands in socrata-test-common
+  import scala.language.implicitConversions
+  class AssertionJSON(actual: => String) {
+      def shouldBeJSON(expected: String) = {
+        val actualObj = Try(parse(actual)).recover({ case jpe: JsonParseException =>
+                          fail(s"""Unable to parse actual value "$actual" as JSON""", jpe)}).get
+        val expectObj = Try(parse(expected)).recover({ case jpe: JsonParseException =>
+                          fail(s"""Unable to parse expected value "$expected" as JSON""", jpe)}).get
+
+        withClue(
+          "\nTextual actual:\n\n" + pretty(render(actualObj)) + "\n\n\n" +
+          "Textual expected:\n\n" + pretty(render(expectObj)) + "\n\n")
+          { actualObj should be (expectObj) }
+      }
+  }
+  implicit def convertJSONAssertion(j: => String): AssertionJSON = new AssertionJSON(j)
+
 
   implicit val execContext = scala.concurrent.ExecutionContext.global
